apps/api/pyproject.toml
[project]
name = "internal-api"
version = "0.1.0"
description = "FastAPI service for internal management/audit/reporting platform"
requires-python = ">=3.14,<3.15"
dependencies = [
  "alembic>=1.16.5",
  "asyncpg>=0.30.0",
  "celery>=5.5.3",
  "fastapi>=0.118.0",
  "greenlet>=3.2.4",
  "PyJWT>=2.10.1",
  "pydantic-settings>=2.11.0",
  "py-common",
  "py-domain",
  "psycopg[binary]>=3.2.10",
  "python-multipart>=0.0.20",
  "redis>=6.4.0",
  "sqlalchemy>=2.0.43",
  "uvicorn>=0.37.0"
]

[tool.uv.sources]
py-common = { workspace = true }
py-domain = { workspace = true }

[project.scripts]
api = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
